+++ regression/run/clibrary/python/test.py
@@ -53,6 +53,15 @@ class Tutorial(unittest.TestCase):
     def testfunction4a(self):
         self.assertEqual('dogcat', clibrary.Function4a('dog', 'cat'))
 
+    def testReturnOneName(self):
+        name1 = clibrary.returnOneName()
+        self.assertEqual("bill", name1)
+
+    def testReturnTwoNames(self):
+        name1, name2 = clibrary.returnTwoNames()
+        self.assertEqual("tom", name1)
+        self.assertEqual("frank", name2)
+
     def testImpliedLen(self):
         rv_int = clibrary.ImpliedLen("bird  ")
         self.assertEqual(6, rv_int)