+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added external attribute to use with function pointers.
   Uses EXTERNAL statement instead of abstract interface to allow any function
   to be used as the dummy argument since the interface is not preserved.
+- Added charlen attribute to use with 'char *arg+intent(out)' argument.
+  Used to tell the Python wrapper the length of the char argument.
 
 ### Fixed
 - C++ function arguments which pass a class by value.